refactor: restructure folders into core/, infra/, internal/
