+++ .gitignore
@@ -1,3 +1,46 @@
+# Node modules
 node_modules/
+
+# Build output
 dist/
+build/
+
+# TypeScript cache
+*.tsbuildinfo
+
+# Test coverage
+coverage/
+
+# Logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+
+# Environment variables
 .env
+.env.local
+.env.*.local
+
+# OS files
+.DS_Store
+Thumbs.db
+
+# IDE config
+.vscode/
+.idea/
+
+# Jest cache
+.jest/
+
+# Swagger generated files (if any)
+swagger.json
+swagger.yaml
+
+# Uploaded files (if applicable in dev)
+uploads/
+
+# Temp files
+tmp/
+temp/
+*.tmp
