+++ package.json
@@ -6,7 +6,7 @@
   "type": "module",
   "types": "./dist/index.d.ts",
   "scripts": {
-    "dev": "ts-node ./src/lib/app.ts",
+    "dev": "ts-node ./src/internal/app.ts",
     "lint": "eslint '**/*.{ts,tsx}'",
     "build": "eslint '**/*.{ts,tsx}' --ignore-pattern '**/__tests__/**' && tsc && tsc-alias",
     "test": "NODE_OPTIONS='--experimental-vm-modules' jest"
@@ -79,5 +79,11 @@
     "ts-node": "^10.0.0",
     "tsc-alias": "^1.8.11",
     "typescript": "^4.0.0"
+  },
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "types": "./dist/index.d.ts"
+    }
   }
 }
