+++ src/core/auth.ts
