+++ src/infra/cron.ts
@@ -1,5 +1,5 @@
 import cron from 'node-cron';
-import { enduranceEmitter } from './emitter.js';
+import { enduranceEmitter } from '../core/emitter';
 
 interface ScheduledTask {
   stop: () => void;
